cloud_functions/climateiq_merge_scenario_predictions_cf/requirements.txt
black==24.4.2
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.10.1
deprecation==2.1.0
flake8==7.0.0
Flask==3.0.3
functions-framework==3.7.0
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
gunicorn==22.0.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
mypy==1.10.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pyflakes==3.2.0
pytest==8.2.2
requests==2.32.3
rsa==4.9
typing_extensions==4.12.2
urllib3==2.2.2
watchdog==4.0.1
Werkzeug==3.0.3
